Get operations using separate func
Having that in separate block provides a possibility to document it by
adding a name to this block. It also should improve possiblity to
unittest it.

If operations are provided, there is no need to verify if view is class
or callable based, so func in this case will immediatelly return the
value.